+++ CHANGELOG.md
@@ -5,8 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
+
 ### Added
 - Alphabet module now supports `NT_STRICT_UNKNOWN` (`ATGCN`)
+
+### Changed
+- GenBank parser was overhauled
+- GenBank parsing now errs on the side of warnings instead of exceptions when encountering bad data
+- Hybrid GenBank parsing now handles duplicate locus tags by shunting them to sorted mode
+
 
 ## [0.13.1] 2022-2-04
 ### Fixed